Disabled dependency to Mono.Posix library, and added latest revision of SharpZipLib to External folder. Corrected cast error in Extensions.cs

+++ Sharpen/Sharpen/Extensions.cs
@@ -608,7 +608,7 @@ namespace Sharpen
 		public static int GetTotalInFixed (this Inflater inf)
 		{
 			if (inf.TotalIn > 0)
-				return inf.TotalIn + 4;
+				return (int)(inf.TotalIn + 4);
 			else
 				return 0;
 		}
